Fix to decimal fault in GeoPointUtil

+++ src/nu/placebo/whatsup/util/GeoPointUtil.java
@@ -21,10 +21,10 @@ public class GeoPointUtil {
 	public static double[] convertAreaToDoubles(GeoPoint a, GeoPoint b){
 		double[] result = new double[4];
 		
-		result[0] = Math.min(a.getLatitudeE6(), b.getLatitudeE6()) /1000000;
-		result[1] = Math.min(a.getLongitudeE6(), b.getLongitudeE6()) /1000000;
-		result[2] = Math.max(a.getLatitudeE6(), b.getLatitudeE6()) /1000000;
-		result[3] = Math.max(a.getLongitudeE6(), b.getLongitudeE6()) /1000000;
+		result[0] = ((double)Math.min(a.getLatitudeE6(), b.getLatitudeE6())) /1000000;
+		result[1] = ((double)Math.min(a.getLongitudeE6(), b.getLongitudeE6())) /1000000;
+		result[2] = ((double)Math.max(a.getLatitudeE6(), b.getLatitudeE6())) /1000000;
+		result[3] = ((double)Math.max(a.getLongitudeE6(), b.getLongitudeE6())) /1000000;
 		
 		return result;
 	}
